+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## NEXT RELEASE
+
+- Fixes a bug where `false` booleans in the GitHub Actions configuration would still pass through truthy. This is because GitHub Actions passes all env vars to the action as a string. We simply check for false strings and coerce them to False bools now (closes #16)
+
 ## v0.12.0 (2022-08-12)
 
 - Adds support for OS/arch targets such as Linux and Darwin, AMD64 and ARM64 (closes #9 & #14)